+++ CHANGELOG.md
@@ -41,6 +41,30 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - **MPX-2230:** update web3 to latest version ([8984335](https://github.com/rarible/sdk/commit/8984335118b979e54388fa294818f68dd3e73d0a))
 - webpack dev server ([9dd2887](https://github.com/rarible/sdk/commit/9dd288726b293a74cfff7f1977098334a6d18b85))
 
+## [0.13.68-fix.68](https://github.com/rarible/sdk/compare/v0.13.68-fix.67...v0.13.68-fix.68) (2024-10-28)
+
+### Features
+
+- add api key form on example ([#679](https://github.com/rarible/sdk/issues/679)) ([e12beb9](https://github.com/rarible/sdk/commit/e12beb9057faf632f0101d6d600cbb1a08b25349))
+
+## [0.13.68-fix.67](https://github.com/rarible/sdk/compare/v0.13.68-fix.66...v0.13.68-fix.67) (2024-10-22)
+
+### Features
+
+- some fixes for eclipse ([#678](https://github.com/rarible/sdk/issues/678)) ([c2cc6c6](https://github.com/rarible/sdk/commit/c2cc6c694fce43697689254d48b89c1bb7788fcd))
+
+## [0.13.68-fix.66](https://github.com/rarible/sdk/compare/v0.13.68-fix.65...v0.13.68-fix.66) (2024-10-22)
+
+### Bug Fixes
+
+- **PLTF-659:** hex data odd length strings looksrare ([#676](https://github.com/rarible/sdk/issues/676)) ([0e6960c](https://github.com/rarible/sdk/commit/0e6960c8814396eaaa33601a81d202a2f0f8fcf7))
+
+## [0.13.68-fix.65](https://github.com/rarible/sdk/compare/v0.13.68-fix.64...v0.13.68-fix.65) (2024-10-21)
+
+### Features
+
+- support eclipse native currency balance ([#673](https://github.com/rarible/sdk/issues/673)) ([d782b59](https://github.com/rarible/sdk/commit/d782b5961448f18a773f52a04607faf161455be9))
+
 ## [0.13.68-fix.64](https://github.com/rarible/sdk/compare/v0.13.68-fix.63...v0.13.68-fix.64) (2024-10-14)
 
 ### Features