+++ packages/eclipse-sdk/CHANGELOG.md
@@ -9,6 +9,26 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - **PLTF-710:** web3 v4 integration provider with new types ([#675](https://github.com/rarible/sdk/issues/675)) ([e7ce3e5](https://github.com/rarible/sdk/commit/e7ce3e549c38fbf2e66f45bf3402926d3b0a1e07))
 
+## [0.13.68-fix.68](https://github.com/rarible/sdk/compare/v0.13.68-fix.67...v0.13.68-fix.68) (2024-10-28)
+
+### Features
+
+- add api key form on example ([#679](https://github.com/rarible/sdk/issues/679)) ([e12beb9](https://github.com/rarible/sdk/commit/e12beb9057faf632f0101d6d600cbb1a08b25349))
+
+## [0.13.68-fix.67](https://github.com/rarible/sdk/compare/v0.13.68-fix.66...v0.13.68-fix.67) (2024-10-22)
+
+### Features
+
+- some fixes for eclipse ([#678](https://github.com/rarible/sdk/issues/678)) ([c2cc6c6](https://github.com/rarible/sdk/commit/c2cc6c694fce43697689254d48b89c1bb7788fcd))
+
+## [0.13.68-fix.66](https://github.com/rarible/sdk/compare/v0.13.68-fix.65...v0.13.68-fix.66) (2024-10-22)
+
+**Note:** Version bump only for package @rarible/eclipse-sdk
+
+## [0.13.68-fix.65](https://github.com/rarible/sdk/compare/v0.13.68-fix.64...v0.13.68-fix.65) (2024-10-21)
+
+**Note:** Version bump only for package @rarible/eclipse-sdk
+
 ## [0.13.68-fix.64](https://github.com/rarible/sdk/compare/v0.13.68-fix.63...v0.13.68-fix.64) (2024-10-14)
 
 ### Features