+++ packages/sdk/CHANGELOG.md
@@ -55,6 +55,26 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - **MPX-2230:** update web3 to latest version ([75c1913](https://github.com/rarible/sdk/commit/75c19137cd770d05862005d4d75a6038ab30cc14))
 - **MPX-2230:** update web3 to latest version ([8984335](https://github.com/rarible/sdk/commit/8984335118b979e54388fa294818f68dd3e73d0a))
 
+## [0.13.68-fix.68](https://github.com/rarible/sdk/compare/v0.13.68-fix.67...v0.13.68-fix.68) (2024-10-28)
+
+### Features
+
+- add api key form on example ([#679](https://github.com/rarible/sdk/issues/679)) ([e12beb9](https://github.com/rarible/sdk/commit/e12beb9057faf632f0101d6d600cbb1a08b25349))
+
+## [0.13.68-fix.67](https://github.com/rarible/sdk/compare/v0.13.68-fix.66...v0.13.68-fix.67) (2024-10-22)
+
+**Note:** Version bump only for package @rarible/sdk
+
+## [0.13.68-fix.66](https://github.com/rarible/sdk/compare/v0.13.68-fix.65...v0.13.68-fix.66) (2024-10-22)
+
+**Note:** Version bump only for package @rarible/sdk
+
+## [0.13.68-fix.65](https://github.com/rarible/sdk/compare/v0.13.68-fix.64...v0.13.68-fix.65) (2024-10-21)
+
+### Features
+
+- support eclipse native currency balance ([#673](https://github.com/rarible/sdk/issues/673)) ([d782b59](https://github.com/rarible/sdk/commit/d782b5961448f18a773f52a04607faf161455be9))
+
 ## [0.13.68-fix.64](https://github.com/rarible/sdk/compare/v0.13.68-fix.63...v0.13.68-fix.64) (2024-10-14)
 
 **Note:** Version bump only for package @rarible/sdk