+++ packages/transaction/CHANGELOG.md
@@ -47,6 +47,24 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @rarible/sdk-transaction
 
+## [0.13.68-fix.68](https://github.com/rarible/sdk/compare/v0.13.68-fix.67...v0.13.68-fix.68) (2024-10-28)
+
+### Features
+
+- add api key form on example ([#679](https://github.com/rarible/sdk/issues/679)) ([e12beb9](https://github.com/rarible/sdk/commit/e12beb9057faf632f0101d6d600cbb1a08b25349))
+
+## [0.13.68-fix.67](https://github.com/rarible/sdk/compare/v0.13.68-fix.66...v0.13.68-fix.67) (2024-10-22)
+
+**Note:** Version bump only for package @rarible/sdk-transaction
+
+## [0.13.68-fix.66](https://github.com/rarible/sdk/compare/v0.13.68-fix.65...v0.13.68-fix.66) (2024-10-22)
+
+**Note:** Version bump only for package @rarible/sdk-transaction
+
+## [0.13.68-fix.65](https://github.com/rarible/sdk/compare/v0.13.68-fix.64...v0.13.68-fix.65) (2024-10-21)
+
+**Note:** Version bump only for package @rarible/sdk-transaction
+
 ## [0.13.68-fix.64](https://github.com/rarible/sdk/compare/v0.13.68-fix.63...v0.13.68-fix.64) (2024-10-14)
 
 **Note:** Version bump only for package @rarible/sdk-transaction